+++ src/main/java/ru/mail/polis/dao/storage/utils/BytesUtils.java
@@ -1,5 +1,6 @@
 package ru.mail.polis.dao.storage.utils;
 
+import com.google.common.base.Charsets;
 import org.jetbrains.annotations.NotNull;
 
 import java.nio.ByteBuffer;
@@ -31,6 +32,14 @@ public final class BytesUtils {
     }
 
     /**
+     *  Return key as ByteBuffer.
+     * @param id  is key witch we should be convert to ByteBuffer
+     */
+    public static ByteBuffer keyByteBuffer(@NotNull final String id) {
+        return ByteBuffer.wrap(id.getBytes(Charsets.UTF_8));
+    }
+
+    /**
      * Transform ByteBuffer to array of bytes.
      * @param buffer is ByteBuffer which will be transform
      */
@@ -39,5 +48,24 @@ public final class BytesUtils {
         final byte[] array = new byte[copy.remaining()];
         copy.get(array);
         return array;
+    }
+
+    /**
+     * Get bytes as massive from kee string.
+     * @param id is id from response
+     */
+    public static byte[] getBytesFromKey(@NotNull final String id) {
+        return id.getBytes(Charsets.UTF_8);
+    }
+
+    /**
+     * Return ByteBuffer as body.
+     * @param value is value to bytes massive
+     */
+    public static byte[] body(@NotNull final ByteBuffer value) {
+        final ByteBuffer duplicate = value.duplicate();
+        final byte[] body = new byte[duplicate.remaining()];
+        duplicate.get(body);
+        return body;
     }
 }
