+++ src/main/java/ru/mail/polis/service/topology/ConsistingHashTopology.java
@@ -3,6 +3,7 @@ package ru.mail.polis.service.topology;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.mail.polis.dao.storage.utils.BytesUtils;
 import ru.mail.polis.service.topology.hash.HashFunction;
 import ru.mail.polis.service.topology.node.ServiceNode;
 import ru.mail.polis.service.topology.node.VirtualNode;
@@ -141,9 +142,7 @@ public final class ConsistingHashTopology implements Topology<ServiceNode> {
             lock.readLock().lock();
             try {
                 messageDigest.reset();
-                final ByteBuffer duplicate = key.duplicate();
-                final byte[] bytes = new byte[duplicate.remaining()];
-                key.get(bytes);
+                final byte[] bytes = BytesUtils.body(key);
                 messageDigest.update(bytes);
                 final byte[] digest = messageDigest.digest();
                 long hash = 0;
